Remove rxjava1 dependencies and add contributor file
