+++ store2store/CONTRIBUTORS.md
@@ -1,0 +1,15 @@
+## store2store contributors
+In contributing order
+
+* **[Thibaud Giovannetti](https://github.com/tspoke)**
+
+  * Author and maintainer
+  * Most features development
+  
+
+* **[Hoang Hiep](https://github.com/hoanghiephui)**
+
+  * Port the library to RxJava2
+  
+
+
