+++ store2store/build.gradle
@@ -5,7 +5,7 @@ group='com.github.playmoweb'
 
 android {
     compileSdkVersion 26
-    buildToolsVersion "25.0.3"
+    buildToolsVersion "26.0.1"
 
     defaultConfig {
         minSdkVersion 16
@@ -29,7 +29,7 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile "io.reactivex:rxandroid:1.2.1"
+
     compile "io.reactivex.rxjava2:rxandroid:2.0.1"
     testCompile 'junit:junit:4.12'
 }
