+++ store2store/src/androidTest/java/com/playmoweb/store2store/mock/MemoryService.java
@@ -6,7 +6,7 @@ import com.playmoweb.store2store.utils.SortingMode;
 
 import java.util.List;
 
-import rx.Observable;
+import io.reactivex.Observable;
 
 /**
  * Memory service for tests purposes
