+++ store2store/src/main/java/com/playmoweb/store2store/service/AbstractService.java
@@ -8,18 +8,21 @@ import com.playmoweb.store2store.utils.SortingMode;
 
 import java.util.List;
 
-import rx.Observable;
-import rx.Subscription;
-import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Func1;
-import rx.schedulers.Schedulers;
-import rx.subscriptions.CompositeSubscription;
+import io.reactivex.Observable;
+import io.reactivex.ObservableSource;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.functions.Function;
+import io.reactivex.schedulers.Schedulers;
 
 /**
  * This abstract service hide basic implementation of CRUD operations combined with a storage (dao)
  * @author  Thibaud Giovannetti
  * @by      Playmoweb
  * @date    08/02/2017.
+ *
+ * @updated hoanghiep
+ * @date    28/07/2017
  */
 public abstract class AbstractService<T> implements IService<T> {
     /**
@@ -35,10 +38,11 @@ public abstract class AbstractService<T> implements IService<T> {
     /**
      * A local subscription to handle local observers
      */
-    protected final CompositeSubscription subscriptions = new CompositeSubscription();
+    protected final CompositeDisposable compositeDisposable = new CompositeDisposable();
 
     /**
      * Public constructor
+     *
      * @param clazz
      */
     public AbstractService(Class<T> clazz, IStoreDao<T> storage) {
@@ -57,20 +61,20 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<List<T>> getAll(final Filter filter, final SortingMode sortingMode, CustomObserver<List<T>> otherSubscriber) {
         Observable<List<T>> observable = getAll(filter, sortingMode)
-                .flatMap(new Func1<List<T>, Observable<List<T>>>() {
+                .flatMap(new Function<List<T>, ObservableSource<List<T>>>() {
                     @Override
-                    public Observable<List<T>> call(final List<T> ts) {
-                        return storage.deleteAll().map(new Func1<Void, List<T>>() {
+                    public ObservableSource<List<T>> apply(final List<T> ts) throws Exception {
+                        return storage.deleteAll().map(new Function<Void, List<T>>() {
                             @Override
-                            public List<T> call(Void aVoid) {
+                            public List<T> apply(Void aVoid) throws Exception {
                                 return ts;
                             }
                         });
                     }
                 })
-                .flatMap(new Func1<List<T>, Observable<List<T>>>() {
+                .flatMap(new Function<List<T>, ObservableSource<List<T>>>() {
                     @Override
-                    public Observable<List<T>> call(List<T> ts) {
+                    public ObservableSource<List<T>> apply(List<T> ts) throws Exception {
                         storage.insertOrUpdate(ts);
                         return storage.getAll(filter, sortingMode);
                     }
@@ -85,15 +89,14 @@ public abstract class AbstractService<T> implements IService<T> {
     }
 
     @Override
-    public final Observable<T> getById(final int id, final CustomObserver<T> otherSubscriber) {
+    public Observable<T> getById(int id, CustomObserver<T> otherSubscriber) {
         Observable<T> observable = getById(id)
-                .flatMap(new Func1<T, Observable<T>>() {
+                .flatMap(new Function<T, ObservableSource<T>>() {
                     @Override
-                    public Observable<T> call(T itemFromAsync) {
+                    public ObservableSource<T> apply(T itemFromAsync) throws Exception {
                         return storage.insertOrUpdate(itemFromAsync);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.getOne(new Filter("id", id), null);
     }
@@ -101,13 +104,12 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<T> getOne(Filter filter, SortingMode sortingMode, CustomObserver<T> otherSubscriber) {
         Observable<T> observable = getOne(filter, sortingMode)
-                .flatMap(new Func1<T, Observable<T>>() {
+                .flatMap(new Function<T, ObservableSource<T>>() {
                     @Override
-                    public Observable<T> call(T itemFromAsync) {
+                    public ObservableSource<T> apply(T itemFromAsync) throws Exception {
                         return storage.insertOrUpdate(itemFromAsync);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.getOne(filter, sortingMode);
     }
@@ -115,24 +117,24 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<T> insert(final T object, CustomObserver<T> otherSubscriber) {
         Observable<T> observable = insert(object)
-                .onErrorResumeNext(new Func1<Throwable, Observable<T>>() {
+                .onErrorResumeNext(new Function<Throwable, ObservableSource<T>>() {
                     @Override
-                    public Observable<T> call(final Throwable throwable) {
-                        return storage.delete(object).flatMap(new Func1<Void, Observable<T>>() {
-                            @Override
-                            public Observable<T> call(Void aVoid) {
-                                return Observable.error(throwable);
-                            }
-                        });
+                    public ObservableSource<T> apply(final Throwable throwable) throws Exception {
+                        return storage.delete(object)
+                                .flatMap(new Function<Void, ObservableSource<T>>() {
+                                    @Override
+                                    public ObservableSource<T> apply(Void aVoid) throws Exception {
+                                        return Observable.error(throwable);
+                                    }
+                                });
                     }
                 })
-                .flatMap(new Func1<T, Observable<T>>() {
+                .flatMap(new Function<T, ObservableSource<T>>() {
                     @Override
-                    public Observable<T> call(T item) {
+                    public ObservableSource<T> apply(T item) throws Exception {
                         return storage.insertOrUpdate(item);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.insertOrUpdate(object);
     }
@@ -140,52 +142,49 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<List<T>> insert(final List<T> objects, CustomObserver<List<T>> otherSubscriber) {
         Observable<List<T>> observable = insert(objects)
-                .onErrorResumeNext(new Func1<Throwable, Observable<List<T>>>() {
+                .onErrorResumeNext(new Function<Throwable, ObservableSource<List<T>>>() {
                     @Override
-                    public Observable<List<T>> call(final Throwable throwable) {
-                        return storage.delete(objects).flatMap(new Func1<Void, Observable<List<T>>>() {
+                    public ObservableSource<List<T>> apply(final Throwable throwable) throws Exception {
+                        return storage.delete(objects).flatMap(new Function<Void, ObservableSource<List<T>>>() {
                             @Override
-                            public Observable<List<T>> call(Void aVoid) {
+                            public ObservableSource<List<T>> apply(Void aVoid) throws Exception {
                                 return Observable.error(throwable);
                             }
                         });
                     }
                 })
-                .flatMap(new Func1<List<T>, Observable<List<T>>>() {
+                .flatMap(new Function<List<T>, ObservableSource<List<T>>>() {
                     @Override
-                    public Observable<List<T>> call(List<T> items) {
-                        return storage.insertOrUpdate(items);
+                    public ObservableSource<List<T>> apply(List<T> ts) throws Exception {
+                        return storage.insertOrUpdate(ts);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.insertOrUpdate(objects);
     }
 
     @Override
-    public Observable<T> update(final T object, CustomObserver<T> otherSubscriber) {
+    public Observable<T> update(T object, CustomObserver<T> otherSubscriber) {
         Observable<T> observable = update(object)
-                .flatMap(new Func1<T, Observable<T>>() {
+                .flatMap(new Function<T, ObservableSource<T>>() {
                     @Override
-                    public Observable<T> call(T item) {
+                    public ObservableSource<T> apply(T item) throws Exception {
                         return storage.insertOrUpdate(item);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return Observable.just(object);
     }
 
     @Override
-    public Observable<List<T>> update(final List<T> objects, CustomObserver<List<T>> otherSubscriber) {
+    public Observable<List<T>> update(List<T> objects, CustomObserver<List<T>> otherSubscriber) {
         Observable<List<T>> observable = update(objects)
-                .flatMap(new Func1<List<T>, Observable<List<T>>>() {
+                .flatMap(new Function<List<T>, ObservableSource<List<T>>>() {
                     @Override
-                    public Observable<List<T>> call(List<T> items) {
-                        return storage.insertOrUpdate(items);
+                    public ObservableSource<List<T>> apply(List<T> ts) throws Exception {
+                        return storage.insertOrUpdate(ts);
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.insertOrUpdate(objects);
     }
@@ -193,18 +192,18 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<Void> delete(final T object, CustomObserver<Void> otherSubscriber) {
         Observable<Void> observable = delete(object)
-                .onErrorResumeNext(new Func1<Throwable, Observable<Void>>() {
+                .onErrorResumeNext(new Function<Throwable, ObservableSource<Void>>() {
                     @Override
-                    public Observable<Void> call(final Throwable throwable) {
-                        return storage.insertOrUpdate(object).flatMap(new Func1<T, Observable<Void>>() {
-                            @Override
-                            public Observable<Void> call(T item) {
-                                return Observable.error(throwable);
-                            }
-                        });
+                    public ObservableSource<Void> apply(final Throwable throwable) throws Exception {
+                        return storage.insertOrUpdate(object)
+                                .flatMap(new Function<T, ObservableSource<Void>>() {
+                                    @Override
+                                    public ObservableSource<Void> apply(T item) throws Exception {
+                                        return Observable.error(throwable);
+                                    }
+                                });
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.delete(object);
     }
@@ -212,18 +211,18 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<Void> delete(final List<T> objects, CustomObserver<Void> otherSubscriber) {
         Observable<Void> observable = delete(objects)
-                .onErrorResumeNext(new Func1<Throwable, Observable<Void>>() {
+                .onErrorResumeNext(new Function<Throwable, ObservableSource<Void>>() {
                     @Override
-                    public Observable<Void> call(final Throwable throwable) {
-                        return storage.insertOrUpdate(objects).flatMap(new Func1<List<T>, Observable<Void>>() {
-                            @Override
-                            public Observable<Void> call(List<T> items) {
-                                return Observable.error(throwable);
-                            }
-                        });
+                    public ObservableSource<Void> apply(final Throwable throwable) throws Exception {
+                        return storage.insertOrUpdate(objects)
+                                .flatMap(new Function<List<T>, ObservableSource<Void>>() {
+                                    @Override
+                                    public ObservableSource<Void> apply(List<T> ts) throws Exception {
+                                        return Observable.error(throwable);
+                                    }
+                                });
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.delete(objects);
     }
@@ -231,24 +230,18 @@ public abstract class AbstractService<T> implements IService<T> {
     @Override
     public Observable<Void> deleteAll(CustomObserver<Void> otherSubscriber) {
         Observable<Void> observable = deleteAll()
-                .onErrorResumeNext(new Func1<Throwable, Observable<Void>>() {
+                .onErrorResumeNext(new Function<Throwable, ObservableSource<Void>>() {
                     @Override
-                    public Observable<Void> call(final Throwable throwable) {
-                        // rollback storage state
-                        return getAll(null, null).flatMap(new Func1<List<T>, Observable<Void>>() {
-                            @Override
-                            public Observable<Void> call(List<T> ts) {
-                                return storage.insertOrUpdate(ts).flatMap(new Func1<List<T>, Observable<Void>>() {
+                    public ObservableSource<Void> apply(final Throwable throwable) throws Exception {
+                        return getAll(null, null)
+                                .flatMap(new Function<List<T>, ObservableSource<Void>>() {
                                     @Override
-                                    public Observable<Void> call(List<T> ts) {
+                                    public ObservableSource<Void> apply(List<T> ts) throws Exception {
                                         return Observable.error(throwable);
                                     }
                                 });
-                            }
-                        });
                     }
                 });
-
         subscribeNonNullObserver(observable, otherSubscriber);
         return storage.deleteAll();
     }
@@ -258,16 +251,16 @@ public abstract class AbstractService<T> implements IService<T> {
      *
      * @note The defaults schedulers are io() for subscribeOn and Android.mainThread() for observeOn.
      * @param observable
-     * @param otherSubscriber
+     * @param observer
      * @param <S>
      */
-    private <S> void subscribeNonNullObserver(final Observable<S> observable, final CustomObserver<S> otherSubscriber) {
-        if(observable != null && otherSubscriber != null) {
-            final Subscription s = observable
-                    .observeOn(AndroidSchedulers.mainThread())
+    private <S> void subscribeNonNullObserver(final Observable<S> observable, final CustomObserver<S> observer) {
+        if (observable != null && observer != null) {
+            observable
                     .subscribeOn(Schedulers.io())
-                    .subscribe(new SimpleObserver<>(otherSubscriber));
-            subscriptions.add(s);
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(new SimpleObserver<S>(observer));
+            compositeDisposable.add(observable.subscribe());
         }
     }
