+++ store2store/src/main/java/com/playmoweb/store2store/service/IService.java
@@ -5,14 +5,16 @@ import com.playmoweb.store2store.utils.Filter;
 import com.playmoweb.store2store.utils.SortingMode;
 
 import java.util.List;
-
-import rx.Observable;
+import io.reactivex.Observable;
 
 /**
  * Service interface
  * @author  Thibaud Giovannetti
  * @by      Playmoweb
- * @date    08/02/2017.
+ * @date    08/02/2017
+ *
+ * @updated hoanghiep
+ * @date    28/07/2017
  */
 public interface IService<T> {
     /**
